Allowing any root dir for tests.

+++ org.knime.testing/src/org/knime/testing/core/AnalyzeLogFile.java
@@ -190,7 +190,7 @@ public class AnalyzeLogFile {
 
         int posTests = m_numOfTestsRun - m_numOfFailingTests;
         int posRate =
-                (int)Math.round((posTests * 100.0) / (double)m_numOfTestsRun);
+                (int)Math.round((posTests * 100.0) / m_numOfTestsRun);
 
         summary.write("Regression run on " + m_startTime + CRLF);
         summary.write("Tests run: " + m_numOfTestsRun + ", failing: "
